Refactored the code into appropriate packages so that manitainability is better. Now development can be done easier on the individual components
