+++ main_screen/__init__.py
@@ -1,0 +1,2 @@
+from .LabClickableFrame import LabClickableFrame
+from .GithubClickableFrame import GithubClickableFrame
