+++ protein_screen/GlobalObject.py
